pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentweave"
version = "1.0.0"
description = "AgentWeave SDK - Python library for building cross-cloud AI agents with hardened security by default"
authors = [
    {name = "High Velocity Solutions LLC", email = "info@hvs.solutions"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
keywords = ["agent", "spiffe", "security", "a2a", "ai", "opa", "mtls"]

dependencies = [
    "spiffe>=0.2.2",
    "spiffe-tls>=0.1.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "opentelemetry-api>=1.24",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-exporter-otlp>=1.24",
    "prometheus-client>=0.20",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.5",
]

[project.scripts]
agentweave = "agentweave.cli.main:cli"

[project.urls]
Homepage = "https://github.com/hvs-solutions/agentweave"
Documentation = "https://agentweave.readthedocs.io"
Repository = "https://github.com/hvs-solutions/agentweave"
"Bug Tracker" = "https://github.com/hvs-solutions/agentweave/issues"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["agentweave*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
